Exercise1.6 updates
Updated the index.html file to include an empty unordered list and gave it a class to make it selectable in the .js file .

Created a variable inside the forEach's loop function block

Assigned the variable the ul element in the html file.

Created an li element

Create a button element (

Set the button element's innerText to be the Pokémon's name.

Added a class to the button using the classList.add method.

Created a CSS rule that targets the button's class .

Gave it a custom button style .

Append the button to the list item as its child.

Append the list item to the unordered list as its child.

Create a function inside the IIFE and name it addListItem().

Used addListItem()’s parameter to set the inner text.

Added the addListItem() function to the object that is returning from the IIFE.

Named the key addListItem().

Called the addListItem() and passed the Pokémon returned in the forEach's loop function block to the addListItem() function call.

Styled the “index.html” page and made it responsive on all common screen sizes. You wan